+++ src/pydidas_plugins/proc_plugins/convert_to_d_spacing.py
@@ -38,6 +38,7 @@ from pydidas.core import (
 )
 from pydidas.core.constants import PROC_PLUGIN_INTEGRATED
 from pydidas.core.generic_params import GENERIC_PARAMS_METADATA
+from pydidas.core.utils.scattering_geometry import convert_polar_to_d_spacing
 from pydidas.plugins import ProcPlugin
 
 
@@ -65,8 +66,6 @@ class ConvertToDSpacing(ProcPlugin):
         self._allowed_ax_choices = _AX_CHOICES
 
     def pre_execute(self):
-        self._lambda = self._EXP.get_param_value("xray_wavelength")
-        self._detector_dist = self._EXP.get_param_value("detector_dist")
         self._config["ax_index"] = None
         self._config["new_range"] = None
         self._config["ax_indices"] = None
@@ -131,31 +130,27 @@ class ConvertToDSpacing(ProcPlugin):
         _axis = self._config["ax_index"]
         _slicer = [slice(None)] * data.ndim
         _range = data.axis_ranges[_axis].copy()
-        match data.axis_labels[_axis]:
-            case "Q":
-                if data.axis_units[_axis] == "nm^-1":
-                    _range = _range / 10
-                _range = (2 * np.pi) / _range
-            case "r":
-                _range = self._lambda / (
-                    2 * np.sin(np.arctan(_range / (self._detector_dist * 1e3)) / 2)
-                )
-            case "2theta":
-                if data.axis_units[_axis] == "deg":
-                    _range = np.radians(_range)
-                _range = self._lambda / (2 * np.sin(_range / 2))
-            case "d-spacing":
-                if data.axis_units[_axis] == "nm":
-                    _range = _range * 10
-        if self.get_param_value("d_spacing_unit") == "nm":
-            _range /= 10
-        _valid = np.isfinite(_range)
-        if not np.all(_valid):
-            _range = _range[_valid]
-        if data.axis_labels[_axis] == "d-spacing":
-            _slicer[_axis] = np.where(_valid)[0]
+        _label_in = data.axis_labels[_axis]
+        _unit_in = data.axis_units[_axis]
+        _unit_out = self.get_param_value("d_spacing_unit")
+
+        if _label_in == "d-spacing":
+            if _unit_in == "nm" and _unit_out in ["A", "Angstrom"]:
+                _range *= 10
+            elif _unit_in == "A" and _unit_out == "nm":
+                _range /= 10
             self._config["new_range"] = _range
         else:
+            _range = convert_polar_to_d_spacing(
+                _range,
+                f"{_label_in} / {_unit_in}",
+                _unit_out,
+                self._EXP.xray_wavelength_in_m,
+                self._EXP.detector_dist_in_m,
+            )
+            _valid = np.isfinite(_range)
+            if not np.all(_valid):
+                _range = _range[_valid]
             _slicer[_axis] = np.where(_valid)[0][::-1]
             self._config["new_range"] = _range[::-1]
         self._slicer = tuple(_slicer)